memory.cの追加　ファイルの分割のため
