+++ bootpack.h
@@ -131,3 +131,24 @@ void inthandler2c(int *esp);
 void enable_mouse(struct MOUSE_DEC *mdec);
 int mouse_decode(struct MOUSE_DEC *mdec, unsigned char data);
 extern struct FIFO8 mousefifo;
+
+/* memory.c */
+#define MEM_FREES 4090 /* 約32GB */
+#define MEM_ADDR 0x003c0000
+
+/* 空き情報 */
+struct FREEINFO
+{
+    unsigned int addr, size;
+};
+/* メモリ管理 */
+struct MEMMAN
+{
+    int frees, maxfrees, lostsize, losts;
+    struct FREEINFO free[MEM_FREES];
+};
+unsigned int memtest(unsigned int start, unsigned int end);
+void memman_init(struct MEMMAN *man);
+unsigned int memman_total(struct MEMMAN *man);
+unsigned int memman_alloc(struct MEMMAN *man, unsigned int size);
+int memman_free(struct MEMMAN *man, unsigned int addr, unsigned int size);
